introduce compute_related_cves() on FIPSDataset
